Do math, but don't do boring math
For division it isn't possible to divide by zero and it isn't
interesting to divide 1 or zero by anything, nor divide by 1.

Same for multiplication, We don't need same result as one of the
inputs, or zero as result.

I am not quite sure if -1 could be filtered away too, so I haven't.